Move /pkg/hardware to /pkg/providers/tinkerbell/hardware
