+++ pkg/providers/tinkerbell/hardware/config.go
