+++ pkg/providers/tinkerbell/hardware/csvparser.go
