+++ pkg/providers/tinkerbell/hardware/jsonparser.go
