+++ pkg/providers/tinkerbell/hardware/jsonparser_test.go
@@ -4,7 +4,7 @@ import (
 	"testing"
 
 	"github.com/aws/eks-anywhere/internal/test"
-	"github.com/aws/eks-anywhere/pkg/hardware"
+	"github.com/aws/eks-anywhere/pkg/providers/tinkerbell/hardware"
 )
 
 func TestNewJsonParserSuccess(t *testing.T) {
