+++ pkg/providers/tinkerbell/hardware/testdata/expected_hardware_json.json
