+++ pkg/providers/tinkerbell/hardware/testdata/expected_hardware_yaml.yaml
