+++ pkg/providers/tinkerbell/hardware/testdata/invalid_hardware_missing_header.csv
