+++ pkg/providers/tinkerbell/hardware/testdata/valid_hardware.csv
