+++ pkg/providers/tinkerbell/hardware/yamlparser.go
