+++ pkg/providers/tinkerbell/validator_helper_test.go
@@ -1,6 +1,6 @@
 package tinkerbell
 
-import "github.com/aws/eks-anywhere/pkg/hardware"
+import "github.com/aws/eks-anywhere/pkg/providers/tinkerbell/hardware"
 
 // SetValidatorHardwareConfig sets v internal hardwareConfig member to cfg.
 func SetValidatorHardwareConfig(v *Validator, cfg hardware.HardwareConfig) {
